requirements.txt
# Runtime (Streamlit Cloud)
streamlit>=1.36
pandas>=2.2
numpy>=1.26
pyarrow>=16
openpyxl>=3.1
scipy>=1.13
scikit-learn>=1.4
statsmodels>=0.14
prophet>=1.1.5
plotly>=5.22
pydeck>=0.9
folium>=0.17
streamlit-folium>=0.20
geopandas>=1.0
shapely>=2.0
pyproj>=3.6
fiona>=1.9
rtree>=1.2
unidecode>=1.3
PyYAML>=6.0
xlsxwriter>=3.2
